Improve tests; we dont need path hook for uv and add a fixture to conftest
