+++ rocq-remote-agent/tests/conftest.py
@@ -1,13 +1,19 @@
-import sys
-from pathlib import Path
+from __future__ import annotations
+
+import pytest
 
 
-def pytest_configure() -> None:
-    """
-    Ensure `src/` is importable when running tests without installing the
-    package.
-    """
-    repo_root = Path(__file__).resolve().parents[1]
-    src = repo_root / "src"
-    if src.is_dir():
-        sys.path.insert(0, str(src))
+class DummyWS:
+    async def send(self, _message: str | bytes) -> None:  # pragma: no cover
+        return None
+
+    async def recv(self) -> str | bytes:  # pragma: no cover
+        return b""
+
+    async def close(self) -> None:  # pragma: no cover
+        return None
+
+
+@pytest.fixture
+def dummy_ws() -> DummyWS:
+    return DummyWS()
